fix according to given remarks
