+++ ewm-service/src/main/java/ru/practicum/utils/CustomPageRequest.java
@@ -1,4 +1,4 @@
-package ru.practicum;
+package ru.practicum.utils;
 
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
